[docs] Refactor out PackageInstallation component and use across MDX docs
