+++ docs/src/docs/form/use-form.mdx
@@ -11,23 +11,14 @@ source: 'mantine-form/src'
 license: MIT
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { FormDemos } from '@mantine/demos';
 
 ## Installation
 
 `@mantine/form` package does not depend on any other libraries, you can use it with or without `@mantine/core` inputs:
 
-Install with npm:
-
-```bash
-npm install @mantine/form
-```
-
-Install with yarn:
-
-```bash
-yarn add @mantine/form
-```
+<PackageInstallation packages={['@mantine/form']} />
 
 ## Usage
