+++ docs/src/docs/others/dropzone.mdx
@@ -23,23 +23,14 @@ import {
   MS_EXCEL_MIME_TYPE,
   MS_POWERPOINT_MIME_TYPE,
 } from '@mantine/dropzone';
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { DropzoneDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/dropzone
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/dropzone
-```
+<PackageInstallation packages={['@mantine/dropzone']} />
 
 ## Usage
