+++ docs/src/docs/others/modals.mdx
@@ -13,23 +13,14 @@ installation: '@mantine/modals'
 license: MIT
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { ModalsDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/modals
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/modals
-```
+<PackageInstallation packages={['@mantine/modals']} />
 
 ## Setup ModalsProvider
