+++ docs/src/docs/others/notifications.mdx
@@ -13,23 +13,14 @@ installation: '@mantine/notifications'
 license: MIT
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { NotificationDemos, NotificationsDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/notifications
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/notifications
-```
+<PackageInstallation packages={['@mantine/notifications']} />
 
 ## Demo
