+++ docs/src/docs/others/nprogress.mdx
@@ -13,23 +13,14 @@ installation: '@mantine/nprogress'
 license: MIT
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { NProgressDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/nprogress
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/nprogress
-```
+<PackageInstallation packages={['@mantine/nprogress']} />
 
 ## Setup NavigationProgress
