+++ docs/src/docs/others/prism.mdx
@@ -14,23 +14,14 @@ license: MIT
 styles: ['Prism', 'PrismTabs']
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { PrismDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/prism
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/prism
-```
+<PackageInstallation packages={['@mantine/prism']} />
 
 ## Usage
