+++ docs/src/docs/others/spotlight.mdx
@@ -14,23 +14,14 @@ license: MIT
 styles: ['SpotlightProvider']
 ---
 
+import { PackageInstallation } from '../../components/PackageInstallation/PackageInstallation';
 import { SpotlightDemos } from '@mantine/demos';
 
 ## Installation
 
 Package depends on `@mantine/core` and `@mantine/hooks`.
 
-Install with yarn:
-
-```bash
-yarn add @mantine/spotlight
-```
-
-Install with npm:
-
-```bash
-npm install @mantine/spotlight
-```
+<PackageInstallation packages={['@mantine/spotlight']} />
 
 ## Usage
